feat: refactoring with security improvements

+++ Models/Http_Response/ProdutosResponse.cs
@@ -5,7 +5,7 @@ namespace Models.HttpResponse
         public long Id {get; set;}
         public required string Nome_Produto {get;set;}   
         public required string Descricao {get;set;}
-        public required int Quantidade {get;set;}
+        public required long Quantidade {get;set;}
         public required double Preco {get;set;}
     }
 }
